Switch to using `0` for false, and `1` for true.
This is a good idea because, folding by or(`|`) can now be implemented
as to find the maximum value, without changing its behaviour for
booleans. In the same way and(`&`) folding can now be implemented as
finding the minimum value.

TODO: Make fold shortcircuit on encountering `0` on `&` and `1` on `|`,
if the expression is confirmed to be a set of booleans.